support expire time editing

+++ core/controller/cluster/models_status.go
@@ -8,7 +8,7 @@ type GetClusterStatusResponse struct {
 	RunningTask       *RunningTask       `json:"runningTask" yaml:"runningTask"`
 	LatestPipelinerun *LatestPipelinerun `json:"latestPipelinerun,omitempty"`
 	ClusterStatus     interface{}        `json:"clusterStatus,omitempty" yaml:"clusterStatus,omitempty"`
-	TTLSeconds        *float64           `json:"ttlSeconds,omitempty" yaml:"ttlSeconds,omitempty"`
+	TTLSeconds        *uint              `json:"ttlSeconds,omitempty" yaml:"ttlSeconds,omitempty"`
 }
 
 // RunningTask 最近一次在执行中的Task
